requirements.txt
python-telegram-bot==20.8
aiosqlite==0.19.0